PS_Graph_DB/pyproject.toml
[project]
name = "graph-schema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.10",
]
